requirements.txt
requests
beautifulsoup4
playwright>=1.40.0
supabase>=1.0.0
python-dotenv>=0.19.0
python-dateutil>=2.8.0
